+++ app/src/main/java/com/example/dessinpartage/metier/Line.java
@@ -4,9 +4,9 @@ import android.graphics.Canvas;
 
 public class Line extends Shape{
 
-    private int x2, y2;
+    private float x2, y2;
 
-    public Line(int x, int y, int x2, int y2, int col) {
+    public Line(float x, float y, float x2, float y2, int col) {
         super(x, y, col);
         this.x2 = x2;
         this.y2 = y2;
