+++ app/src/main/java/com/example/dessinpartage/metier/Rectangle.java
@@ -5,9 +5,9 @@ import android.graphics.Paint;
 
 public class Rectangle extends Shape{
 
-    private int x2, y2;
+    private float x2, y2;
 
-    public Rectangle(int x, int y, int x2, int y2, int col, boolean plein) {
+    public Rectangle(float x, float y, float x2, float y2, int col, boolean plein) {
         super(x, y, col);
         this.x2 = x2;
         this.y2 = y2;
